refactor: rename _dynamics to _src
