+++ src/galax/dynamics/_src/funcs.py
