+++ src/galax/dynamics/_src/integrate/funcs.py
@@ -15,7 +15,7 @@ import galax.coordinates as gc
 import galax.potential as gp
 import galax.typing as gt
 from .core import Integrator
-from galax.dynamics._dynamics.orbit import Orbit
+from galax.dynamics._src.orbit import Orbit
 
 ##############################################################################
