+++ src/galax/dynamics/_src/mockstream/_moving.py
