+++ src/galax/dynamics/_src/mockstream/core.py
