+++ src/galax/dynamics/_src/mockstream/df/base.py
@@ -18,8 +18,8 @@ import galax.coordinates as gc
 import galax.potential as gp
 import galax.typing as gt
 from .progenitor import ConstantMassProtenitor, ProgenitorMassCallable
-from galax.dynamics._dynamics.mockstream.core import MockStreamArm
-from galax.dynamics._dynamics.orbit import Orbit
+from galax.dynamics._src.mockstream.core import MockStreamArm
+from galax.dynamics._src.orbit import Orbit
 
 Carry: TypeAlias = tuple[gt.LengthVec3, gt.SpeedVec3, gt.LengthVec3, gt.SpeedVec3]
