+++ src/galax/dynamics/_src/mockstream/df/chen24.py
@@ -17,7 +17,7 @@ import quaxed.numpy as jnp
 import galax.potential as gp
 import galax.typing as gt
 from .base import AbstractStreamDF
-from galax.dynamics._dynamics.funcs import specific_angular_momentum, tidal_radius
+from galax.dynamics._src.funcs import specific_angular_momentum, tidal_radius
 
 # ============================================================
 # Constants
