+++ src/galax/dynamics/_src/mockstream/df/fardal15.py
@@ -16,7 +16,7 @@ import quaxed.numpy as jnp
 import galax.potential as gp
 import galax.typing as gt
 from .base import AbstractStreamDF
-from galax.dynamics._dynamics.funcs import _orbital_angular_frequency, tidal_radius
+from galax.dynamics._src.funcs import _orbital_angular_frequency, tidal_radius
 
 # ============================================================
 # Constants
