+++ src/galax/dynamics/_src/mockstream/df/progenitor.py
