+++ src/galax/dynamics/_src/mockstream/mockstream_generator.py
@@ -19,9 +19,9 @@ import galax.typing as gt
 from .core import MockStream, MockStreamArm
 from .df import AbstractStreamDF, ProgenitorMassCallable
 from .utils import cond_reverse
-from galax.dynamics._dynamics.integrate.core import Integrator
-from galax.dynamics._dynamics.integrate.funcs import _default_integrator, evaluate_orbit
-from galax.dynamics._dynamics.orbit import Orbit
+from galax.dynamics._src.integrate.core import Integrator
+from galax.dynamics._src.integrate.funcs import _default_integrator, evaluate_orbit
+from galax.dynamics._src.orbit import Orbit
 from galax.potential import AbstractPotentialBase
 
 Carry: TypeAlias = tuple[gt.IntScalar, gt.VecN, gt.VecN]
