+++ src/galax/dynamics/_src/mockstream/utils.py
