+++ src/galax/dynamics/integrate.py
@@ -1,7 +1,7 @@
 """:mod:`galax.dynamics.integrate`."""
 
-from ._dynamics import integrate
-from ._dynamics.integrate import *  # noqa: F403
+from ._src import integrate
+from ._src.integrate import *  # noqa: F403
 
 __all__: list[str] = []
 __all__ += integrate.__all__
