+++ src/galax/dynamics/mockstream.py
@@ -1,7 +1,7 @@
 """:mod:`galax.dynamics.mockstream`."""
 
-from ._dynamics import mockstream
-from ._dynamics.mockstream import *  # noqa: F403
+from ._src import mockstream
+from ._src.mockstream import *  # noqa: F403
 
 __all__: list[str] = []
 __all__ += mockstream.__all__
